fix(project): translating pages and content to english
