+++ posts/2022-01-26-use-effect-vs-use-layout-effect.mdx
@@ -1,10 +1,10 @@
 ---
-title: Olá mundo jan!
-subtitle: Subtítulo jan!
+title: Hello World!
+subtitle: I'm amazing!
 description: Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old.
 hero_image: "/assets/uploads/cover/light.png"
 date: 2020-01-01 12:16:25
-author: Eu mesmo
+author: My self
 category: 'Front-end'
 series: ''
 tags: ['javascript']
