+++ src/utils/algolia_queries.js
@@ -10,7 +10,7 @@ const postQuery = `{
         }
         frontmatter {
           date_timestamp: date
-          date(locale: "pt-br", formatString: "DD [de] MMMM [de] YYYY")
+          date(formatString: "MMMM, DD, YYYY")
           description
           title
           subtitle
